Upgrading to .NET 2.0
git-svn-id: https://alchemi.svn.sourceforge.net/svnroot/alchemi/trunk@157 9c0b43c4-e927-0410-a2d5-cf972382494e

+++ alchemi-devel/examples/PiCalculator/PiCalculator/PiCalcGridThread.cs
@@ -40,7 +40,7 @@ namespace Alchemi.Examples.PiCalculator
             StringBuilder temp = new StringBuilder();
 
             Plouffe_Bellard pb = new Plouffe_Bellard();
-            for (int i = 0; i <= Math.Ceiling(_NumDigits / 9); i++)
+            for (int i = 0; i <= Math.Ceiling((double)_NumDigits / 9); i++)
             {
                 temp.Append(pb.CalculatePiDigits(_StartDigitNum + (i * 9)));
             }
